Extract trait for VariantAware implementation
