+++ plugin/src/main/groovy/com/novoda/staticanalysis/internal/CodeQualityConfigurator.groovy
@@ -9,20 +9,18 @@ import org.gradle.api.Task
 import org.gradle.api.plugins.quality.CodeQualityExtension
 import org.gradle.api.tasks.SourceTask
 
-abstract class CodeQualityConfigurator<T extends SourceTask, E extends CodeQualityExtension> implements Configurator {
+abstract class CodeQualityConfigurator<T extends SourceTask, E extends CodeQualityExtension> implements Configurator, VariantAware {
 
     protected final Project project
     protected final Violations violations
     protected final Task evaluateViolations
     protected final SourceFilter sourceFilter
-    protected Closure<Boolean> includeVariantsFilter
 
     protected CodeQualityConfigurator(Project project, Violations violations, Task evaluateViolations) {
         this.project = project
         this.violations = violations
         this.evaluateViolations = evaluateViolations
         this.sourceFilter = new SourceFilter(project)
-        this.includeVariantsFilter = { true }
     }
 
     @Override
@@ -38,13 +36,13 @@ abstract class CodeQualityConfigurator<T extends SourceTask, E extends CodeQuali
             }
             project.plugins.withId('com.android.application') {
                 project.afterEvaluate {
-                    configureAndroidProject(allApplicationVariants.matching { includeVariantsFilter(it) })
+                    configureAndroidProject(filteredApplicationVariants)
                     configureToolTasks()
                 }
             }
             project.plugins.withId('com.android.library') {
                 project.afterEvaluate {
-                    configureAndroidProject(allLibraryVariants.matching { includeVariantsFilter(it) })
+                    configureAndroidProject(filteredLibraryVariants)
                     configureToolTasks()
                 }
             }
@@ -57,27 +55,11 @@ abstract class CodeQualityConfigurator<T extends SourceTask, E extends CodeQuali
         }
     }
 
-    protected NamedDomainObjectSet<Object> getAllApplicationVariants() {
-        getAllVariants(project.android.applicationVariants)
-    }
-
     protected void configureToolTasks() {
         project.tasks.withType(taskClass) { task ->
             task.group = 'verification'
             configureReportEvaluation(task, violations)
         }
-    }
-
-    protected NamedDomainObjectSet<Object> getAllLibraryVariants() {
-        getAllVariants(project.android.libraryVariants)
-    }
-
-    private NamedDomainObjectSet<Object> getAllVariants(variants1) {
-        NamedDomainObjectSet<Object> variants = project.container(Object)
-        variants.addAll(variants1)
-        variants.addAll(project.android.testVariants)
-        variants.addAll(project.android.unitTestVariants)
-        return variants
     }
 
     protected abstract String getToolName()
